fix rm

+++ bin/glue-setup.sh
@@ -16,11 +16,14 @@ GLUE_PY_FILES="$ROOT_DIR/PyGlue.zip"
 export PYTHONPATH="$GLUE_PY_FILES:$PYTHONPATH"
 
 # Run mvn copy-dependencies target to get the Glue dependencies locally
-# mvn -f $ROOT_DIR/pom.xml -DoutputDirectory=$ROOT_DIR/jarsv1 dependency:copy-dependencies
+mvn -f $ROOT_DIR/pom.xml -DoutputDirectory=$ROOT_DIR/jarsv1 dependency:copy-dependencies
 
 export SPARK_CONF_DIR=${ROOT_DIR}/conf
 mkdir $SPARK_CONF_DIR
-rm $SPARK_CONF_DIR/spark-defaults.conf
+
+if [ -e $SPARK_CONF_DIR/spark-defaults.conf ]; then
+  rm $SPARK_CONF_DIR/spark-defaults.conf
+fi
 # Generate spark-defaults.conf
 echo "spark.driver.extraClassPath $GLUE_JARS_DIR/*" >> $SPARK_CONF_DIR/spark-defaults.conf
 echo "spark.executor.extraClassPath $GLUE_JARS_DIR/*" >> $SPARK_CONF_DIR/spark-defaults.conf
